Setting up branch for cmd dev
