+++ minipedia.py
@@ -3,6 +3,7 @@ import json
 import requests
 import re
 import unidecode
+import argparse
 
 charlist = ['a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z',' ','.',',','(',')'," "]
 
@@ -10,7 +11,15 @@ pages=[[],[],[]]
 
 symboldict = {'@' : 'at','!' : '.','&' : 'and','=' : 'equals','+' : 'plus', '-' : 'minus', '0' : 'zero', '1' : 'one', '2' : 'two', '3' : 'three', '4' : 'four', '5' : 'five', '6' : 'six', '7' : 'seven', '8' : 'eight', '9' : 'nine'}
 words = ["zero", "one", "two", "three", "four", "five", "six", "seven", "eight", "nine", "ten"]
-    
+
+iscapital = False
+istitle = True
+
+pages=[]
+cont=""
+gaplimit=""
+filename=""
+
 #Turns 8bit binary into 5bit
 def bitify(inputs):
     binry = ""
@@ -88,3 +97,65 @@ def symboltoplaintext(text3):
     for word, initial in symboldict.items():
         text3 = text3.replace(word.lower(), initial)
     return text3
+
+def debitify(filename):
+    file = open(filename+".5bf", "rb")
+    byte = file.read(1) #add while byte:
+    fullnum = ""
+    fullstr = ""
+    while byte:
+        checknum = "{0:b}".format(int.from_bytes(byte, "little")).zfill(8) #reads byte as integer
+        fullnum = fullnum + checknum
+        if (len(fullnum) > 39):
+            fullnum2 = int(fullnum, 2)
+            for x2 in range(8):
+                fullstr = fullstr + bitcheck(readbits40(fullnum2,x2*5,5))
+            fullnum = ""
+        byte = file.read(1) #new byte 
+    return fullstr
+
+def readbits40(var,start,amount):
+    var = var >> (40-(start+amount)) & (int((40-amount)*"0"+amount*"1",2)) #reads amount amount of bits from position start starting from 0
+    return var
+
+def bitcheck(fivebit):
+    global iscapital
+    global istitle
+    match fivebit: #decodes integer to character
+        case 0:
+            return " "
+        case 27:
+            return "."
+        case 28:
+            return ","
+        case 29:
+            return "("
+        case 30:
+            return ")"  
+        case 31:
+            if (istitle == True):
+                istitle = False
+                return "\n\n"
+            iscapital = True
+            return ""               
+        case _:
+            if (iscapital == False):
+                return chr(fivebit+96) 
+            else:
+                iscapital = False
+                return chr(fivebit+64)
+    return binry 
+
+def encoder(cont):
+    pages=wikigraball(cont,gaplimit)
+    for i in range(len(pages[0])):
+        writefile(addmultititle(pages[0][i],pages[1][i]),filename)
+    cont=pages[2][0]
+    return cont
+
+def decode():
+    filename = input("Enter article name: ")
+    print(debitify(filename))
+
+parser=argparse.ArgumentParser()
+parser.parse_args()
